@@ -37,8 +37,8 @@
     "@enact/core": "^2.2.9",
     "prop-types": "^15.6.0",
     "ramda": "^0.24.1",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
     "xhr": "^2.4.1"
   }
 }
