@@ -52,16 +52,6 @@ const GridListImageItemBase = kind({
 
 	propTypes: /** @lends moonstone/GridListImageItem.GridListImageItemBase.prototype */ {
 		/**
-		 * The voice control intent.
-		 *
-		 * @type {String}
-		 * @default 'Select'
-		 * @memberof moonstone/GridListImageItem.GridListImageItemBase.prototype
-		 * @public
-		 */
-		'data-webos-voice-intent': PropTypes.string,
-
-		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
 		 * corresponding internal Elements and states of this component.
 		 *
@@ -77,6 +67,16 @@ const GridListImageItemBase = kind({
 		 * @public
 		 */
 		css: PropTypes.object,
+
+		/**
+		 * The voice control intent.
+		 *
+		 * @type {String}
+		 * @default 'Select'
+		 * @memberof moonstone/GridListImageItem.GridListImageItemBase.prototype
+		 * @public
+		 */
+		'data-webos-voice-intent': PropTypes.string,
 
 		/**
 		 * Placeholder image used while [source]{@link ui/GridListImageItem.GridListImageItem#source}
