@@ -68,6 +68,14 @@ const IncrementSliderBase = kind({
 
 	propTypes: /** @lends moonstone/IncrementSlider.IncrementSliderBase.prototype */ {
 		/**
+		 * Sets the knob to selected state and allows it to move via 5-way controls.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		active: PropTypes.bool,
+
+		/**
 		 * Prevents read out of both the slider and the increment and decrement
 		 * buttons.
 		 *
@@ -89,32 +97,6 @@ const IncrementSliderBase = kind({
 		'aria-valuetext': PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
 
 		/**
-		 * Disables voice control.
-		 *
-		 * @type {Boolean}
-		 * @memberof moonstone/IncrementSlider.IncrementSliderBase.prototype
-		 * @public
-		 */
-		'data-webos-voice-disabled': PropTypes.bool,
-
-		/**
-		 * The `data-webos-voice-group-label` for the IconButton of IncrementSlider.
-		 *
-		 * @type {String}
-		 * @memberof moonstone/IncrementSlider.IncrementSliderBase.prototype
-		 * @public
-		 */
-		'data-webos-voice-group-label': PropTypes.string,
-
-		/**
-		 * Sets the knob to selected state and allows it to move via 5-way controls.
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		active: PropTypes.bool,
-
-		/**
 		 * Background progress, as a proportion between `0` and `1`.
 		 *
 		 * @type {Number}
@@ -131,6 +113,24 @@ const IncrementSliderBase = kind({
 		 * @private
 		 */
 		css: PropTypes.object,
+
+		/**
+		 * Disables voice control.
+		 *
+		 * @type {Boolean}
+		 * @memberof moonstone/IncrementSlider.IncrementSliderBase.prototype
+		 * @public
+		 */
+		'data-webos-voice-disabled': PropTypes.bool,
+
+		/**
+		 * The `data-webos-voice-group-label` for the IconButton of IncrementSlider.
+		 *
+		 * @type {String}
+		 * @memberof moonstone/IncrementSlider.IncrementSliderBase.prototype
+		 * @public
+		 */
+		'data-webos-voice-group-label': PropTypes.string,
 
 		/**
 		* Sets the hint string read when focusing the decrement button.
