@@ -78,6 +78,15 @@ class ScrollableBaseNative extends Component {
 		childRenderer: PropTypes.func.isRequired,
 
 		/**
+		 * Animate while scrolling
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @private
+		 */
+		animate: PropTypes.bool,
+
+		/**
 		 * This is set to `true` by SpotlightContainerDecorator
 		 *
 		 * @type {Boolean}
@@ -102,15 +111,6 @@ class ScrollableBaseNative extends Component {
 		 * @private
 		 */
 		'data-spotlight-id': PropTypes.string,
-
-		/**
-		 * Animate while scrolling
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @private
-		 */
-		animate: PropTypes.bool,
 
 		/**
 		 * Direction of the list or the scroller.
